+++ apps/hyperlocalise-web/src/workflows/translation-job.ts
@@ -1,56 +1,16 @@
 import { fetch as workflowFetch, getWorkflowMetadata } from "workflow";
-
-import { env } from "@/lib/env";
 import type { TranslationJobQueuedEventData } from "@/lib/workflow/types";
-
-function getInternalApiUrl(path: string): string {
-  const { url } = getWorkflowMetadata();
-  return `${url}/api/internal/workflow${path}`;
-}
-
-function internalApiHeaders(): Record<string, string> {
-  const headers: Record<string, string> = {
-    "content-type": "application/json",
-  };
-  if (env.WORKFLOW_INTERNAL_SECRET) {
-    headers["x-internal-secret"] = env.WORKFLOW_INTERNAL_SECRET;
-  }
-  return headers;
-}
+import {
+  getInternalApiUrl,
+  internalApiHeaders,
+  type ClaimedTranslationJob,
+  type ClaimTranslationJobResult,
+} from "./internal-api";
 
 type ClaimTranslationJobInput = {
   event: TranslationJobQueuedEventData;
   runId: string;
 };
-
-type ClaimedTranslationJob = {
-  id: string;
-  projectId: string;
-  type: "string" | "file";
-  inputPayload: unknown;
-  workflowRunId: string;
-};
-
-type ClaimTranslationJobResult =
-  | {
-      kind: "claimed";
-      job: ClaimedTranslationJob;
-    }
-  | {
-      kind: "skipped";
-      job: {
-        id: string;
-        projectId: string;
-        type: "string" | "file";
-        status: string;
-        inputPayload: unknown;
-        outcomeKind: string | null;
-        outcomePayload: unknown;
-        lastError: string | null;
-        workflowRunId: string | null;
-        completedAt: string | null;
-      };
-    };
 
 type TranslationJobExecutionResult =
   | {
